headache

+++ main.h
@@ -12,9 +12,9 @@ int _printf(const char *format, ...);
 int _putchar(char f);
 int _puts(const char *str);
 int put_int(int num);
-int binary_put(int n);
-int put_unsigned(unsigned int num);
-int print_octal(unsigned int num);
-int lower_hex(unsigned int num);
-int upper_hex(unsigned int num);
+int binary_put(long int n);
+int put_unsigned(long unsigned int num);
+int print_octal(long unsigned int num);
+int lower_hex(long unsigned int num);
+int upper_hex(long unsigned int num);
 #endif
